feat: match history
Bot now has new matches command to view the 10 most recent matches for a given summoner.
An array of Match objects are stored within Summoner. Each match object contains an exhuastive amount of data regarding the match.
A list of matches with basic data pertaining to the requested summoner are displayed in an embed via the bot.py matches method.
Future features will allow indidual matches with more in depth data to be viewed via a dropdown under this embed.